[datakit] Rename run_source_normalizations.py -> trigger_sources.py
Adds --downloads-only to materialize just normalize_steps[0] per source
(StepRunner still dedupes shared family downloads via output_path).
Drops the hardcoded MARIN_PREFIX=gs://marin-us-central1 — staging region
now comes from the caller's iris --region pin, which auto-exports
MARIN_PREFIX on the worker.
